// src/interfaces/wms/transaction/inbound/tallyDetails_wms.interface.ts
export interface ITallyDetailsWms {
  company_code?: string;
  prin_code?: string;
  job_no?: string;
  container_no?: string;
  prod_code: string;
  prod_attrib_code?: string;
  quantity?: number;
  pda_quantity?: number;
  location_code?: string;
  barcode?: string;
  size_value?: string;
  seq_number?: number;
  user_id?: string;
  user_dt?: Date;
  packdet_no?: number;
  carton_no?: string;
  batch_no?: string;
  prod_exp_date?: Date;
  pallet_id?: string;
  selected?: string;
  allocated?: string;
  pda_qty_puom?: number;
  pda_puom?: string;
  pda_qty_luom?: number;
  pda_luom?: string;
  lot_no?: string;
  prod_mfg_date?: Date;
  site_ind?: string;
  tally_processed?: string;
  putaway_processed?: string;
  target_location?: string;
  receipt_type?: string;
  vessel_name?: string;
  exp_date?: Date;
  mfg_date?: Date;
  uppp?: number;
  origin_country?: string;
  batch_id?: string;
  carton_tally?: string;
  putaway_pda_processed?: string;
  prod_exp_char?: string;
  prod_mfg_char?: string;
  updated_at?: Date;
  updated_by?: string;
  created_by?: string;
  created_at?: Date;
  po_no?: string;
  gross_weight?: number;
  volume?: number;
  shelf_life_days?: number;
  shelf_life_date?: Date;
  prod_name?: string;
  qty_string?: string
}